Fix out-of-memory issue reported by Samsung
After we changed the actual update part to async, it is possible
multiple delta update can be run before the very first update finishes.
Each of those delta update will query CP2 and possible make both
AppSearch and CP2 hold memory until the update finishes.

To avoid that, we should reset the mDeltaUpdatePending after we finished
updating the Person corpus in AppSearch.

In order to handle the change notificaitons from CP2 during the update,
a new flag mHasNotificationDuringDeltaUpdateFromCP2 is introduced and can be set to
true, so later we can check this flag and run an additional delta update
to handle any missing notifications.

Bug: b/243819878
Test: atest ContactsIndexerTests
Change-Id: Icaf81cf7eff6345e0766f22bb5c832dcb2b6d5d7