Makes ContactsIndexerImpl create a local ContactsBatcher in indexContactsFromCursorAsync() for indexing contacts.
This fixes an issue where simultaneous full and delta updates would use
the same contacts batcher and index each others' contacts. Generally
speaking, this isn't an issue since all the contacts get indexed either
way; however, the updated contacts stats that are logged are wrong when
this happens, and we can even get negative values for the number of
failed updates.

Bug: b/290824217
Test: ContactsIndexerUserInstanceTest.java
Change-Id: I7ed0f7e6bbd82700828e2887cb1faa64e7f36aa1