Safeparcelables in AppSearch - VisibilityDocument
Make VisibilityDocument and VisibilityPermissionDocument SafeParcelable,
and remove their inheritance to GenericDocument. GenericDocument will
contain a GenericDocumentParcel in the future, and it is unnecessary to
have this internal SafeParcel for VisibilityDocument. Plus, we have plan
NOT to store visibility documents in IcingLib in the future:
b/298118943.

For now, we still need to convert those two to GenericDocuments until
we completely treat them differently for b/298118943.

The related change in Jetpack: aosp/2733415

Bug: 275629842
Test: atest CtsAppSearchTestCases CtsAppSearchHostTestCases AppSearchServicesTests AppSearchCoreTests ContactsIndexerTests

Change-Id: I17fb1f1a95bef6d690c3ae388555a94675cdb130