Change default byte and time optimize thresholds
In our experiment, we removed the 1 mb bytes optimize threshold and then set the time optimize threshold to 1 week as a safety precaution (from max infinity). The goal was to lessen how often optimize runs but also ensure it still runs at least once in a while.
We saw 10x less daily optimizations though double max storage size at p99 (7mb -> 16mb) and 30% increased latency on optimize calls. This seems to be a good tradeoff as the total time a device may be spending in optimize per day is significantly less.

This cl raises the byte optimize threshold to 10 mb instead of removing it completely as we did in the experiment. The experiment was on only google pixel devices and we can't guarantee the same behavior for all devices. We choose a 10 mb threshold as a safety precaution because this threshold should preserve the results we saw in the experiment (at p99, we had a 9 mb difference in before and after storage sizes).

Summary of experiment stats: http://shortn/_WXmMp29XWF

Bug: 296918757
Test: FrameworkAppSearchConfigTest.java
Change-Id: I40d2f63d2709ecdc3da5e84b6050f464e049dbaf