Fix appsearch can not see the latest value of the flag
This is based on how packages/modules/HealthFitness has made Flags
available to both framework and service. They have two separate targets,
one that is exported and one that is not. THey use the exported one for
tests, and the unexported one for service & framework. In
service/Android.bp, we do have to include the flag library as a library
rather than a static_library as it is already included in the framework
jar, which is included in the service jar.
https://cs.android.com/android/platform/superproject/main/+/main:packages/modules/HealthFitness/flags/Android.bp

Test: log the flag value
Bug: 275592563
Flag: EXEMPT bugfix

Change-Id: I4c5ae9d8934fba868378ae0edfee614dd89bcde7