Fix bug with enterprise contacts test
If there was a pre-existing contact with the same id as the inserted
enterprise profile test contact, the test would grab the lookup key and
photo uri from the pre-existing (wrong) contact and use those to compare
with the enterprise test contact's fields.

The test now loops through the query results to make sure it grabs the
fields from the correct contact.

Bug: 377042909
Test: EnterpriseContactsTest.java
Flag: TEST_ONLY
Change-Id: I93d282acbea64cee2efe68842c0e495971953f9e