Throw in AppSearchManager.executeAppFunction
There is a security issue reported in this code path.
This API was never supposed to be launched and guarded behind a flag.
However, an attacker can still invoke the hidden API and exploit
the bug. Throwing an exception for that case.

In main, we already removed the code, but need to fix this in Android 15
and so this change.


Bug: 392614489
Test: Presubmit
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:b1d2fcc8b39be70b99e783ee90428d4958231bc9)
Merged-In: Ia5bf6a6133aac5f83c5ec062ccd25662bd29a8b0
Change-Id: Ia5bf6a6133aac5f83c5ec062ccd25662bd29a8b0